# ⚡️ Check for empty string using empty string instead of strlen